refactor: 모달창 오류 수정
