+++ src/components/AlertModal.vue
@@ -2,7 +2,7 @@
   <Transition name="toast">
     <div class="toast-message" v-if="show">
       <div class="toast-content">
-        <span class="check-icon">✓</span>
+        <!-- <span class="check-icon">✓</span> -->
         {{ message }}
       </div>
     </div>
@@ -55,13 +55,8 @@ onMounted(() => {
 .toast-content {
   display: flex;
   align-items: center;
-  gap: 8px;
   font-size: 16px;
   font-weight: 500;
-}
-
-.check-icon {
-  font-size: 18px;
 }
 
 /* 토스트 애니메이션 */
